clase medida actualizada

+++ app/app/src/main/java/com/example/lpamvil/sprint0laura/Medida.java
@@ -5,8 +5,8 @@ public class Medida {
     //Atributos de la clase
     private int valor;
     private String fecha;
-    private int latitud;
-    private int longitud;
+    private long latitud;
+    private long longitud;
 
 
     //Getters y Setters de los atributos
@@ -26,7 +26,7 @@ public class Medida {
         this.fecha = fecha;
     }
 
-    public int getLatitud() {
+    public long getLatitud() {
         return latitud;
     }
 
@@ -34,7 +34,7 @@ public class Medida {
         this.latitud = latitud;
     }
 
-    public int getLongitud() {
+    public long getLongitud() {
         return longitud;
     }
 
@@ -48,7 +48,7 @@ public class Medida {
 
 
     //Constructor con los atributos
-    public Medida(int valor, String fecha, int latitud, int longitud) {
+    public Medida(int valor, String fecha, long latitud, long longitud) {
         this.valor = valor;
         this.fecha = fecha;
         this.latitud = latitud;
